rename fixture file
